Last VS 2015 fixes

+++ neo/game/SecurityCamera.cpp
@@ -141,7 +141,7 @@ void idSecurityCamera::Spawn( void ) {
 	}
 
 	negativeSweep = ( sweepAngle < 0 ) ? true : false;
-	sweepAngle = abs( sweepAngle );
+	sweepAngle = abs(static_cast<long>(sweepAngle) );
 
 	scanFovCos = cos( scanFov * idMath::PI / 360.0f );
